requirements.txt
# pip install -r requirements.txt
# === FastAPI 서버 ===
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# === 비동기 처리 & 웹 ===
nest_asyncio==1.6.0
httpx==0.27.0

# === LangChain + 관련 모듈 (버전 조정) ===
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.38
langchainhub>=0.1.15
langgraph>=0.0.30
langsmith>=0.0.77
openai==1.24.1
langchain-google-vertexai>=0.1.3
langchain-qdrant>=0.2.0

# === Embedding & VectorDB ===
sentence-transformers==2.6.1
qdrant-client==1.10.1
tqdm==4.66.2

# === Google Cloud 연동 ===
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.124.0
google-cloud-pubsub==2.21.0

# === 기타 개발 유틸 ===
pydantic==2.7.4
aiohttp==3.9.5
typing-extensions==4.12.0